// controls/richtexteditor/blazor-script/src/editor-manager/base/interface.ts
export interface IHtmlSubCommands {
    /**
     * Specifies the item
     */
    item?: IAdvanceListItem
    /**
     * Specifies the subCommand.
     */
    subCommand: string
    /**
     * Specifies the callBack.
     */
    callBack(args: IHtmlFormatterCallBack): () => void
    /**
     * Specifies the callBack.
     */
    value?: string | Node
    /**
     * Specifies the originalEvent.
     */
    event?: MouseEvent
    /**
     * Specifies the iframe element selector.
     */
    selector?: string
    /**
     * Specifies if the icon click is from dropdown or direct toolbarclick.
     */
    exeValue?: { [key: string]: string }
    enterAction?: string
}
// ...
export interface IHtmlKeyboardEvent {
    /**
     * Specifies the callBack.
     */
    callBack(args?: IHtmlFormatterCallBack): () => void
    /**
     * Specifies the event.
     */
    event: KeyboardEventArgs
    /**
     * Specifies the notifier name.
     */
    name?: string
    /**
     * Specifies the enter key configuration.
     */
    enterAction?: string
    /**
     * Specifies the tab key action in the Rich Text Editor content..
     */
    enableTabKey?: string
    /**
     * Defines tag to be used when enter key is pressed.
     */
    enterKey?: string
    /**
     * Defines tag to be used when shift enter key is pressed.
     */
    shiftEnterKey?: string
}